morning coffee. cleaning up random sandbox warnings

+++ geometric.hpp
@@ -212,7 +212,7 @@ namespace avl
         double x = sr*cp*cy - cr*sp*sy;
         double y = cr*sp*cy + sr*cp*sy;
         double z = cr*cp*sy - sr*sp*cy;
-        return float4(x,y,z,w);
+        return float4(float(x), float(y), float(z), float(w));
     }
     
     inline float3 make_euler_from_quat(float4 q)
@@ -222,9 +222,9 @@ namespace avl
         const double q1 = q.x;
         const double q2 = q.y;
         const double q3 = q.z;
-        e.x = atan2(2*(q0*q1+q2*q3), 1-2*(q1*q1+q2*q2));
-        e.y = asin(2*(q0*q2-q3*q1));
-        e.z = atan2(2*(q0*q3+q1*q2), 1-2*(q2*q2+q3*q3));
+        e.x = float(atan2(2*(q0*q1+q2*q3), 1.0 - 2.0 * (q1*q1+q2*q2)));
+        e.y = float(asin(2*(q0*q2-q3*q1)));
+        e.z = float(atan2(2*(q0*q3+q1*q2), 1.0 - 2.0 *(q2*q2+q3*q3)));
         return e;
     }
